+++ src/sentry/conf/server.py
@@ -1410,7 +1410,7 @@ SENTRY_DEFAULT_INTEGRATIONS = (
     "sentry.integrations.jira_server.JiraServerIntegrationProvider",
     "sentry.integrations.vsts.VstsIntegrationProvider",
     "sentry.integrations.vsts_extension.VstsExtensionIntegrationProvider",
-    "sentry.integrations.pagerduty.PagerDutyIntegrationProvider",
+    "sentry.integrations.pagerduty.integration.PagerDutyIntegrationProvider",
 )
