+++ src/sentry/integrations/pagerduty/__init__.py
@@ -1,10 +1,7 @@
 from __future__ import absolute_import
 
-from sentry.utils.imports import import_submodules
 from sentry.rules import rules
 
 from .notify_action import PagerDutyNotifyServiceAction
-
-import_submodules(globals(), __name__, __path__)
 
 rules.add(PagerDutyNotifyServiceAction)
